# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Orcafrio - Sistema de Orçamentos de Refrigeração
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cliente {
  id: number;
  nome: string;
  telefone: string;
  email?: string | null;
  endereco?: string | null;
  cpfCnpj?: string | null;
  createdAt: string;
}

export interface CreateClienteBody {
  nome: string;
  telefone: string;
  email?: string | null;
  endereco?: string | null;
  cpfCnpj?: string | null;
}

export type ItemOrcamentoCategoria =
  (typeof ItemOrcamentoCategoria)[keyof typeof ItemOrcamentoCategoria];

export const ItemOrcamentoCategoria = {
  manutencao_preventiva: "manutencao_preventiva",
  higienizacao: "higienizacao",
  instalacao: "instalacao",
  troca_compressor: "troca_compressor",
  carga_gas: "carga_gas",
  diagnostico: "diagnostico",
  conserto: "conserto",
  mao_de_obra: "mao_de_obra",
  outros: "outros",
} as const;

export interface ItemOrcamento {
  id: number;
  orcamentoId: number;
  categoria: ItemOrcamentoCategoria;
  descricao: string;
  quantidade: number;
  precoUnitario: number;
  subtotal: number;
}

export type CreateItemBodyCategoria =
  (typeof CreateItemBodyCategoria)[keyof typeof CreateItemBodyCategoria];

export const CreateItemBodyCategoria = {
  manutencao_preventiva: "manutencao_preventiva",
  higienizacao: "higienizacao",
  instalacao: "instalacao",
  troca_compressor: "troca_compressor",
  carga_gas: "carga_gas",
  diagnostico: "diagnostico",
  conserto: "conserto",
  mao_de_obra: "mao_de_obra",
  outros: "outros",
} as const;

export interface CreateItemBody {
  categoria: CreateItemBodyCategoria;
  descricao: string;
  quantidade: number;
  precoUnitario: number;
}

export type OrcamentoStatus =
  (typeof OrcamentoStatus)[keyof typeof OrcamentoStatus];

export const OrcamentoStatus = {
  rascunho: "rascunho",
  enviado: "enviado",
  aprovado: "aprovado",
  recusado: "recusado",
  cancelado: "cancelado",
} as const;

export interface Orcamento {
  id: number;
  numero: string;
  clienteId: number;
  cliente?: Cliente;
  status: OrcamentoStatus;
  prazoExecucao?: string | null;
  validadeOrcamento?: string | null;
  garantia?: string | null;
  condicoesPagamento?: string | null;
  observacoes?: string | null;
  total: number;
  itens: ItemOrcamento[];
  createdAt: string;
  updatedAt: string;
}

export type OrcamentoListItemStatus =
  (typeof OrcamentoListItemStatus)[keyof typeof OrcamentoListItemStatus];

export const OrcamentoListItemStatus = {
  rascunho: "rascunho",
  enviado: "enviado",
  aprovado: "aprovado",
  recusado: "recusado",
  cancelado: "cancelado",
} as const;

export interface OrcamentoListItem {
  id: number;
  numero: string;
  clienteNome: string;
  status: OrcamentoListItemStatus;
  total: number;
  createdAt: string;
}

export interface CreateOrcamentoBody {
  clienteId: number;
  prazoExecucao?: string | null;
  validadeOrcamento?: string | null;
  garantia?: string | null;
  condicoesPagamento?: string | null;
  observacoes?: string | null;
  itens: CreateItemBody[];
}

export interface UpdateOrcamentoBody {
  clienteId?: number;
  prazoExecucao?: string | null;
  validadeOrcamento?: string | null;
  garantia?: string | null;
  condicoesPagamento?: string | null;
  observacoes?: string | null;
  itens?: CreateItemBody[];
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  rascunho: "rascunho",
  enviado: "enviado",
  aprovado: "aprovado",
  recusado: "recusado",
  cancelado: "cancelado",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export type DashboardResumoPorStatus = {
  rascunho: number;
  enviado: number;
  aprovado: number;
  recusado: number;
  cancelado: number;
};

export interface DashboardResumo {
  totalOrcamentos: number;
  totalClientes: number;
  valorTotalAprovado: number;
  valorTotalPendente: number;
  porStatus: DashboardResumoPorStatus;
}

export type ListClientesParams = {
  search?: string;
};

export type ListOrcamentosParams = {
  status?: ListOrcamentosStatus;
  clienteId?: number;
  search?: string;
};

export type ListOrcamentosStatus =
  (typeof ListOrcamentosStatus)[keyof typeof ListOrcamentosStatus];

export const ListOrcamentosStatus = {
  rascunho: "rascunho",
  enviado: "enviado",
  aprovado: "aprovado",
  recusado: "recusado",
  cancelado: "cancelado",
} as const;