# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Orcafrio - Sistema de Orçamentos de Refrigeração
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clientes
    description: Gerenciamento de clientes
  - name: orcamentos
    description: Gerenciamento de orçamentos
  - name: dashboard
    description: Dados do dashboard

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Listar clientes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Criar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "201":
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"

  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Obter cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Atualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "200":
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Deletar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cliente deletado

  /orcamentos:
    get:
      operationId: listOrcamentos
      tags: [orcamentos]
      summary: Listar orçamentos
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [rascunho, enviado, aprovado, recusado, cancelado]
        - name: clienteId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de orçamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrcamentoListItem"
    post:
      operationId: createOrcamento
      tags: [orcamentos]
      summary: Criar orçamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrcamentoBody"
      responses:
        "201":
          description: Orçamento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orcamento"

  /orcamentos/{id}:
    get:
      operationId: getOrcamento
      tags: [orcamentos]
      summary: Obter orçamento completo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Orçamento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orcamento"
    put:
      operationId: updateOrcamento
      tags: [orcamentos]
      summary: Atualizar orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrcamentoBody"
      responses:
        "200":
          description: Orçamento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orcamento"
    delete:
      operationId: deleteOrcamento
      tags: [orcamentos]
      summary: Deletar orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Orçamento deletado

  /orcamentos/{id}/status:
    patch:
      operationId: updateOrcamentoStatus
      tags: [orcamentos]
      summary: Atualizar status do orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orcamento"

  /orcamentos/{id}/itens:
    post:
      operationId: addOrcamentoItem
      tags: [orcamentos]
      summary: Adicionar item ao orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Item adicionado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemOrcamento"

  /orcamentos/{id}/itens/{itemId}:
    put:
      operationId: updateOrcamentoItem
      tags: [orcamentos]
      summary: Atualizar item do orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "200":
          description: Item atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemOrcamento"
    delete:
      operationId: deleteOrcamentoItem
      tags: [orcamentos]
      summary: Remover item do orçamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item removido

  /dashboard/resumo:
    get:
      operationId: getDashboardResumo
      tags: [dashboard]
      summary: Resumo do dashboard
      responses:
        "200":
          description: Resumo com totais e contagens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResumo"

  /dashboard/orcamentos-recentes:
    get:
      operationId: getOrcamentosRecentes
      tags: [dashboard]
      summary: Orçamentos recentes
      responses:
        "200":
          description: Lista dos 5 orçamentos mais recentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrcamentoListItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        cpfCnpj:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - nome
        - telefone
        - createdAt

    CreateClienteBody:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        cpfCnpj:
          type: string
          nullable: true
      required:
        - nome
        - telefone

    ItemOrcamento:
      type: object
      properties:
        id:
          type: integer
        orcamentoId:
          type: integer
        categoria:
          type: string
          enum: [manutencao_preventiva, higienizacao, instalacao, troca_compressor, carga_gas, diagnostico, conserto, mao_de_obra, outros]
        descricao:
          type: string
        quantidade:
          type: number
        precoUnitario:
          type: number
        subtotal:
          type: number
      required:
        - id
        - orcamentoId
        - categoria
        - descricao
        - quantidade
        - precoUnitario
        - subtotal

    CreateItemBody:
      type: object
      properties:
        categoria:
          type: string
          enum: [manutencao_preventiva, higienizacao, instalacao, troca_compressor, carga_gas, diagnostico, conserto, mao_de_obra, outros]
        descricao:
          type: string
        quantidade:
          type: number
        precoUnitario:
          type: number
      required:
        - categoria
        - descricao
        - quantidade
        - precoUnitario

    Orcamento:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteId:
          type: integer
        cliente:
          $ref: "#/components/schemas/Cliente"
        status:
          type: string
          enum: [rascunho, enviado, aprovado, recusado, cancelado]
        prazoExecucao:
          type: string
          nullable: true
        validadeOrcamento:
          type: string
          nullable: true
        garantia:
          type: string
          nullable: true
        condicoesPagamento:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        total:
          type: number
        itens:
          type: array
          items:
            $ref: "#/components/schemas/ItemOrcamento"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - numero
        - clienteId
        - status
        - total
        - itens
        - createdAt
        - updatedAt

    OrcamentoListItem:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteNome:
          type: string
        status:
          type: string
          enum: [rascunho, enviado, aprovado, recusado, cancelado]
        total:
          type: number
        createdAt:
          type: string
      required:
        - id
        - numero
        - clienteNome
        - status
        - total
        - createdAt

    CreateOrcamentoBody:
      type: object
      properties:
        clienteId:
          type: integer
        prazoExecucao:
          type: string
          nullable: true
        validadeOrcamento:
          type: string
          nullable: true
        garantia:
          type: string
          nullable: true
        condicoesPagamento:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        itens:
          type: array
          items:
            $ref: "#/components/schemas/CreateItemBody"
      required:
        - clienteId
        - itens

    UpdateOrcamentoBody:
      type: object
      properties:
        clienteId:
          type: integer
        prazoExecucao:
          type: string
          nullable: true
        validadeOrcamento:
          type: string
          nullable: true
        garantia:
          type: string
          nullable: true
        condicoesPagamento:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        itens:
          type: array
          items:
            $ref: "#/components/schemas/CreateItemBody"

    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [rascunho, enviado, aprovado, recusado, cancelado]
      required:
        - status

    DashboardResumo:
      type: object
      properties:
        totalOrcamentos:
          type: integer
        totalClientes:
          type: integer
        valorTotalAprovado:
          type: number
        valorTotalPendente:
          type: number
        porStatus:
          type: object
          properties:
            rascunho:
              type: integer
            enviado:
              type: integer
            aprovado:
              type: integer
            recusado:
              type: integer
            cancelado:
              type: integer
          required:
            - rascunho
            - enviado
            - aprovado
            - recusado
            - cancelado
      required:
        - totalOrcamentos
        - totalClientes
        - valorTotalAprovado
        - valorTotalPendente
        - porStatus